+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export { cn } from "./cn";
+export * from "./color";
 export * from "tailwind-variants";