+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # mizuhara
 
+## 2.6.0
+
+### Minor Changes
+
+- Export color utils
+
 ## 2.5.0
 
 ### Minor Changes
 
-- add applyBase options to shadcn
+- Add applyBase options to shadcn
 
 ## 2.4.0